ai-knowledge-sprint/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
weaviate-client==3.25.3
sentence-transformers==2.2.2
comet-ml==3.35.3
scikit-learn==1.3.2
numpy==1.26.2
httpx==0.25.1
pydantic==2.5.0
